backend/image_generator/requirements.txt
Flask
tritonclient[http]
numpy
Pillow
googletrans==4.0.0-rc1